split ModuleProvider, use llc 1.6.1
